pyproject.toml
[project]
name = "datasette-load"
version = "0.1a4"
description = "API and UI for bulk loading data into Datasette from a URL"
readme = "README.md"
authors = [{name = "Datasette"}]
license = "Apache-2.0"
classifiers=[
    "Framework :: Datasette"
]
requires-python = ">=3.10"
dependencies = [
    "datasette>=1.0a21"
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project.urls]
Homepage = "https://github.com/datasette/datasette-load"
Changelog = "https://github.com/datasette/datasette-load/releases"
Issues = "https://github.com/datasette/datasette-load/issues"
CI = "https://github.com/datasette/datasette-load/actions"

[project.entry-points.datasette]
load = "datasette_load"

[project.optional-dependencies]
test = ["pytest", "pytest-asyncio", "pytest-httpx"]

[tool.pytest.ini_options]
asyncio_mode = "strict"
asyncio_default_fixture_loop_scope = "function"

[tool.setuptools.package-data]
datasette_load = ["templates/*"]
